Upstream version 2.2.30.2

# gpg: Signature made Πεμ 22 Ιούν 2017 08:52:34 μμ EEST
# gpg:                using RSA key 3E02FD6656295952110BAB99F51B18C720248224
# gpg:                issuer "apoikos@debian.org"
# gpg: Good signature from "Apollon Oikonomopoulos <apoikos@dmesg.gr>" [ultimate]
# gpg:                 aka "Apollon Oikonomopoulos <apoikos@gmail.com>" [ultimate]
# gpg:                 aka "Apollon Oikonomopoulos <apoikos@debian.org>" [ultimate]

+++ src/auth/userdb-dict.c
@@ -1,4 +1,4 @@
-/* Copyright (c) 2013-2016 Dovecot authors, see the included COPYING file */
+/* Copyright (c) 2013-2017 Dovecot authors, see the included COPYING file */
 
 #include "auth-common.h"
 #include "userdb.h"
@@ -22,7 +22,7 @@ struct dict_userdb_iterate_context {
 
 	userdb_callback_t *userdb_callback;
 	const char *key_prefix;
-	unsigned int key_prefix_len;
+	size_t key_prefix_len;
 	struct dict_iterate_context *iter;
 };
