+++ src/lib/str.c
@@ -1,4 +1,4 @@
-/* Copyright (c) 2002-2016 Dovecot authors, see the included COPYING file */
+/* Copyright (c) 2002-2017 Dovecot authors, see the included COPYING file */
 
 #include "lib.h"
 #include "buffer.h"
@@ -106,7 +106,7 @@ void str_vprintfa(string_t *str, const char *fmt, va_list args)
 #define SNPRINTF_INITIAL_EXTRA_SIZE 128
 	va_list args2;
 	char *tmp;
-	unsigned int init_size;
+	size_t init_size;
 	size_t pos = str->used;
 	int ret, ret2;
